requirements.txt
fastapi==0.124.2
uvicorn[standard]==0.38.0
python-jose[cryptography]==3.5.0
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
openstacksdk==3.3.0
structlog==24.1.0
